Moved Interact Functionality
Moved the Interact functionality from the RotatablePuzzleObject to a new InteractablePuzzleObject to help better seperate functionality and keep code more easily readable.

Added a Camera to InteractablePuzzleObject so that when the player interacts with it the camera transitions to a view of RotatablePuzzleObject to help make the puzzle easier to solve.

Fixed the Stepped Rotation functionality for the RotatablePuzzleObject so that you can properly step the rotation.

Adjusted the lights in the puzzle room